+++ test/SampleSerializer.partial.cs
@@ -9,12 +9,12 @@ namespace Pingmint.CodeGen.Json.Test
     {
         public sealed partial class Sample : ICount, IName
         {
-            public int? Count { get; set; }
+            public Int64? Count { get; set; }
             public bool? IsTrue { get; set; }
             public List<bool>? Bools { get; set; }
             public String? Name { get; set; }
-            public List<Int32>? Items { get; set; }
-            public Int32? Id { get; set; }
+            public List<Int64>? Items { get; set; }
+            public Int64? Id { get; set; }
             public Sample? Recursion { get; set; }
             public List<Sample>? Items2 { get; set; }
             public Dictionary<String, String>? Extensions { get; set; }
@@ -26,7 +26,7 @@ namespace Pingmint.CodeGen.Json.Test
     }
     public partial interface ICount
     {
-        int? Count { get; set; }
+        Int64? Count { get; set; }
         bool? IsTrue { get; set; }
         List<bool>? Bools { get; set; }
     }
